kernels: a conv reads its input past 2^31 elements — the channel offset is 64-bit (register 80)
The ladder law, once the plan read the card's live memory (4bd6a5e7), handed
real-esrgan-x8 a 9.6 GB tile at 1536² and four runs of four died with CUDA 700.
Pinned with --mode triton-sequential and CUDA_LAUNCH_BLOCKING=1: the launch of
conv2d_forward_kernel at 1x64x6344x6344 -> 3 channels, the network's last conv.
Its output (241 MB) is under the 4 GiB band-streaming threshold so it ran whole;
its input is 2 576 000 000 elements, and `input_in_feat_stride * inf_offset`
— a loop-derived int32 arange times a 40 247 936-element channel plane — wraps
past channel 53. Register 58's class: the 09-14 promotion widened every offset
born from a program id; this one is born from a loop.

`inf_offset` (conv2d) and `input_c_offset` (conv1d) are int64.

Cell: 64 channels of 6000² fp16 (2 304 000 000 elements), channels 60-63 —
entirely past 2^31 — hold 2.0, the rest 1.0, a 3x3 mean kernel; every interior
pixel must read 1.0625. RED on the kernel as it stood (cudaMemcpy rc=700 at the
first checked call), GREEN after; the 56 existing conv cells pass.

+++ src/neurobrix/kernels/ops/conv1d.py
@@ -67,7 +67,7 @@ def conv1d_forward_kernel(
         c = (lc % BLOCK_CI_COUNT) * BLOCK_CI
         k = lc // BLOCK_CI_COUNT
 
-        input_c_offset = c + tl.arange(0, BLOCK_CI)
+        input_c_offset = (c + tl.arange(0, BLOCK_CI)).to(tl.int64)   # the channel-plane product must not wrap at 2^31 (conv2d, 2026-09-20)
         input_l_offset = k * dilation_l - padding_l + stride_l * out_l_point
 
         curr_input_pointer = (
